Split large CLI command modules
