+++ packages/cli/src/commands/restore.ts
@@ -1,24 +1,20 @@
 import {
-  boot,
   isNodeLevel5ProductSnapshotBundle,
   restore,
   restoreNodeLevel5DeclaredSubset,
   restoreNodeLevel5ProductSnapshot,
   type VmHandle,
 } from "@machinen/runtime";
-import { existsSync, readFileSync, writeFileSync } from "node:fs";
-import { join, resolve } from "node:path";
-
-import type { PortableNodeSnapshotBundle } from "../clean-service/node-adapter.ts";
+import { resolve } from "node:path";
 import {
   cmdRestoreCleanService as restoreCleanServiceBundle,
   shouldRestoreCleanService as shouldRestoreCleanServiceBundle,
 } from "../clean-service/restore.ts";
 import { consumeJsonFlag, emitJson, emitJsonError } from "../args.ts";
 import {
+  deriveBootName,
   guestCpu,
   resolveCliBaseAssets,
-  sha256Bytes,
   resolveOptionalImageOverride,
   type CliBaseAssetPaths,
 } from "../base-assets.ts";
@@ -35,20 +31,13 @@ import {
 } from "./node-level5-shared.ts";
 import { formatElapsed, isQuiet, NoiseFilter, printHeadline, RingBuffer } from "../quiet.ts";
 import { guestConsoleOnLog, runAttachedVmSession, type QuietRunState } from "../session.ts";
+import {
+  cmdRestorePortableNode,
+  shouldPreferVmstateRestore,
+  shouldRestorePortableNode,
+} from "./restore-portable-node.ts";
 
 type ParsedRestoreCommandArgs = ReturnType<typeof parseRestoreArgs>;
-
-function deriveBootName(imageOverride: string | undefined): string {
-  if (!imageOverride) {
-    return "vm";
-  }
-  const base = imageOverride.split("/").pop() ?? imageOverride;
-  return base
-    .replace(/\.tar\.gz$/i, "")
-    .replace(/\.tgz$/i, "")
-    .replace(/\.tar$/i, "")
-    .replace(/\.gz$/i, "");
-}
 
 function cmdRestoreNodeLevel5DeclaredSubset(input: { json: boolean; rest: string[] }): number {
   const options = parseNodeLevel5DeclaredSubsetRestoreArgs(input.rest);
@@ -211,186 +200,6 @@ async function cmdRestoreNodeLevel5ProofComposition(
     }
   }
   return result.exitCode;
-}
-
-function shouldPreferVmstateRestore(snapDir: string): boolean {
-  if (!existsSync(join(snapDir, "state.vmstate"))) {
-    return false;
-  }
-  const manifestPath = join(snapDir, "portable-node.json");
-  if (!existsSync(manifestPath)) {
-    return true;
-  }
-  const manifest = JSON.parse(readFileSync(manifestPath, "utf8")) as PortableNodeSnapshotBundle;
-  return manifest.sourceArch === guestCpu();
-}
-
-function shouldRestorePortableNode(snapDir: string): boolean {
-  const manifestPath = join(snapDir, "portable-node.json");
-  if (!existsSync(manifestPath)) {
-    return false;
-  }
-  const manifest = JSON.parse(readFileSync(manifestPath, "utf8")) as PortableNodeSnapshotBundle;
-  return manifest.sourceArch !== guestCpu() || !existsSync(join(snapDir, "state.vmstate"));
-}
-
-// fallow-ignore-next-line complexity
-async function cmdRestorePortableNode(
-  parsed: ParsedRestoreCommandArgs,
-  snapDir: string,
-  json: boolean,
-): Promise<number> {
-  const started = Date.now();
-  const manifest = JSON.parse(
-    readFileSync(join(snapDir, "portable-node.json"), "utf8"),
-  ) as PortableNodeSnapshotBundle;
-  if (manifest.sourceArch === guestCpu()) {
-    return reportPortableNodeRestoreRefusal(
-      snapDir,
-      json,
-      manifest,
-      "node-target-architecture-mismatch",
-      "portable Node restore requires a destination architecture different from the source architecture; use vmstate restore for same-architecture bundles",
-      started,
-    );
-  }
-  const appTarPath = join(snapDir, manifest.appTar.path);
-  const appTar = readFileSync(appTarPath);
-  if (sha256Bytes(appTar) !== manifest.appTar.sha256) {
-    return reportPortableNodeRestoreRefusal(
-      snapDir,
-      json,
-      manifest,
-      "node-portable-app-digest-mismatch",
-      "portable Node app tarball digest does not match descriptor",
-      started,
-    );
-  }
-  const paths = await resolveCliBaseAssets();
-  const name = parsed.name ?? deriveBootName(snapDir);
-  const vm = await boot({
-    image: paths.defaultImagePath,
-    kernel: paths.kernelPath,
-    dtb: paths.dtbPath,
-    name,
-    detached: true,
-    cmd: ["sleep", "100000"],
-    timeoutMs: undefined,
-  }).catch(handleError);
-  try {
-    await vm.exec(
-      "export DEBIAN_FRONTEND=noninteractive; " +
-        "if ! command -v node >/dev/null 2>&1; then apt-get update && apt-get install -y --no-install-recommends nodejs curl ca-certificates; fi",
-      { execTimeoutMs: 180_000 },
-    );
-    const targetNode = await vm.execRaw("node --version", { execTimeoutMs: 5_000 });
-    if (targetNode.exitCode !== 0 || targetNode.stdout.trim() !== manifest.nodeVersion) {
-      return reportPortableNodeRestoreRefusal(
-        snapDir,
-        json,
-        manifest,
-        "node-source-target-version-mismatch",
-        `source Node ${manifest.nodeVersion} does not match target Node ${targetNode.stdout.trim() || "unavailable"}`,
-        started,
-      );
-    }
-    await vm.writeFile("/tmp/machinen-portable-node-app.tar.gz", appTar);
-    await vm.exec(
-      "rm -rf /opt/machinen-portable-node-app && mkdir -p /opt/machinen-portable-node-app && tar -xzf /tmp/machinen-portable-node-app.tar.gz -C /opt/machinen-portable-node-app",
-    );
-    await vm.exec(startPortableNodeCommand(manifest), { execTimeoutMs: 15_000 });
-    const verify = await vm.execRaw(verifyPortableNodeCommand(manifest), { execTimeoutMs: 30_000 });
-    if (verify.exitCode !== 0) {
-      return reportPortableNodeRestoreRefusal(
-        snapDir,
-        json,
-        manifest,
-        "node-target-verifier-mismatch",
-        verify.stderr || verify.stdout || "target verifier failed",
-        started,
-      );
-    }
-    const summary = portableNodeRestoreSummary(manifest, "completed", started, {
-      migrationCompleted: true,
-      targetVerifierResult: "passed",
-      restoredName: vm.name ?? name,
-    });
-    // fallow-ignore-next-line code-duplication
-    writeFileSync(
-      join(snapDir, "portable-node-restore-summary.json"),
-      `${JSON.stringify(summary, null, 2)}\n`,
-    );
-    if (json) {
-      emitJson({ schema_version: 1, ...summary });
-    } else {
-      process.stderr.write(`restored portable Node as: ${vm.name ?? name} (pid ${vm.pid})\n`);
-    }
-    return 0;
-  } finally {
-    await vm.detach();
-  }
-}
-
-function startPortableNodeCommand(manifest: PortableNodeSnapshotBundle): string {
-  const nodeIndex = manifest.argv.findIndex((arg) => /(^|\/)node(?:$|[0-9.-])/u.test(arg));
-  const nodeArgs = manifest.argv
-    .slice(nodeIndex + 1)
-    .map(shellQuote)
-    .join(" ");
-  return `cd /opt/machinen-portable-node-app && nohup node ${nodeArgs} >/tmp/machinen-portable-node.log 2>&1 &`;
-}
-
-function verifyPortableNodeCommand(manifest: PortableNodeSnapshotBundle): string {
-  return `for i in $(seq 1 80); do got=$(curl -fsS http://127.0.0.1:${manifest.guestPort}/ 2>/dev/null | sha256sum | awk '{print $1}') && test "$got" = ${shellQuote(manifest.verifier.sha256)} && exit 0; sleep 0.25; done; cat /tmp/machinen-portable-node.log 2>/dev/null; exit 1`;
-}
-
-function reportPortableNodeRestoreRefusal(
-  snapDir: string,
-  json: boolean,
-  manifest: PortableNodeSnapshotBundle,
-  code: string,
-  message: string,
-  started: number,
-): number {
-  const summary = portableNodeRestoreSummary(manifest, "refused", started, {
-    migrationCompleted: false,
-    targetVerifierResult: "failed",
-    refusal: { code, message },
-  });
-  // fallow-ignore-next-line code-duplication
-  writeFileSync(
-    join(snapDir, "portable-node-restore-summary.json"),
-    `${JSON.stringify(summary, null, 2)}\n`,
-  );
-  if (json) {
-    emitJson({ schema_version: 1, ...summary });
-  } else {
-    process.stderr.write(`refused portable Node restore: ${code}: ${message}\n`);
-  }
-  return 1;
-}
-
-function portableNodeRestoreSummary(
-  manifest: PortableNodeSnapshotBundle,
-  state: "completed" | "refused",
-  started: number,
-  extra: Record<string, unknown>,
-): Record<string, unknown> {
-  return {
-    kind: "machinen.portable-node-restore-summary",
-    formatVersion: 1,
-    runtime: "node",
-    subset: manifest.subset,
-    state,
-    sourceArch: manifest.sourceArch,
-    targetArch: guestCpu(),
-    elapsedMs: Date.now() - started,
-    sourceIsaEmulationUsed: false,
-    sourceTextReplayAcceptedAsRestore: false,
-    sidecarRuntimeUsed: false,
-    appHooksRequired: false,
-    ...extra,
-  };
 }
 
 function shellQuote(value: string): string {
